Add support for pelias/geocode.earth geocoder
But don't use it by default.

I will probably end up sticking with geocodio, as it is better at
handling addresses with incorrect street types (it fixes it based on
the postal code instead of dropping the street number and seemingly
choosing a random street in the area). Pelias does do a better job at
choosing a point near the front of the building though.